Introduce new Bike objects and SensorConfigurations
The new data model introduces Bike objects that represent a bike with
all its measurable data (head angle, suspension strokes, linkage joint
positions, and the image the latter was measured from), but has nothing
to do with individual sensor setups. The purpose of this is to make bike
data shareable across users.

Sensor configurations are implemented directly in code rather than the
previous approach of code-in-database. This is admittedly less flexible
than the previous solution, but in practice it results in a much cleaner
implementation.

Bikes and sensor configurations are coupled together using Setup objects,
not unlike the previous solution, where Setups consisted of Linkages and
Calibrations.