Fix max rebound/compression velocity not showing
If max rebound/compression is larger than the display threshold on the
velocity histogram (2000 mm/s), its label would not show up. The
original Bokeh implementation had it at the bottm of the plot, and move
to its proper place when said place was scrolled into view. I'm not sure
 how to do this with ScottPlot, so I just put the label to the proper
 place for now.